requirements.txt
numpy>1.18.5
python-telegram-bot==21.3
